feat: 落页 - v0.5.0

+++ server/config.js
@@ -61,7 +61,7 @@ const Dir = {
 
 // -- 初始化文件夹 --
 
-if (fs.existsSync(Dir.temp)) {
+if (fs.existsSync(Dir.temp) && process.env.NODE_ENV === 'development') {
     fs.rmSync(Dir.temp, { recursive: true });
 }
